backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
